move back route struct to http.go
